+++ jason/tests/peer/media.rs
@@ -13,7 +13,7 @@ use medea_jason::{
 };
 use wasm_bindgen_test::*;
 
-use crate::get_test_tracks;
+use crate::get_test_unrequired_tracks;
 
 wasm_bindgen_test_configure!(run_in_browser);
 
@@ -29,7 +29,7 @@ async fn get_test_media_connections(
         tx,
     );
     let (audio_track, video_track) =
-        get_test_tracks(!enabled_audio, !enabled_video);
+        get_test_unrequired_tracks(!enabled_audio, !enabled_video);
     let audio_track_id = audio_track.id;
     let video_track_id = video_track.id;
     media_connections
@@ -48,11 +48,13 @@ async fn get_test_media_connections(
     media_connections
         .get_sender_by_id(audio_track_id)
         .unwrap()
-        .mute_state_transition_to(StableMuteState::from(!enabled_audio));
+        .mute_state_transition_to(StableMuteState::from(!enabled_audio))
+        .unwrap();
     media_connections
         .get_sender_by_id(video_track_id)
         .unwrap()
-        .mute_state_transition_to(StableMuteState::from(!enabled_video));
+        .mute_state_transition_to(StableMuteState::from(!enabled_video))
+        .unwrap();
 
     (media_connections, audio_track_id, video_track_id)
 }
@@ -66,7 +68,7 @@ fn get_stream_request1() {
         Rc::new(RtcPeerConnection::new(Vec::new(), false).unwrap()),
         tx,
     );
-    let (audio_track, video_track) = get_test_tracks(false, false);
+    let (audio_track, video_track) = get_test_unrequired_tracks(false, false);
     media_connections
         .create_tracks(vec![audio_track, video_track])
         .unwrap();
@@ -110,7 +112,9 @@ async fn disable_and_enable_all_tracks_in_media_manager() {
     assert!(!audio_track.is_muted());
     assert!(!video_track.is_muted());
 
-    audio_track.mute_state_transition_to(StableMuteState::Muted);
+    audio_track
+        .mute_state_transition_to(StableMuteState::Muted)
+        .unwrap();
     media_connections
         .update_senders(vec![TrackPatch {
             id: audio_track_id,
@@ -120,7 +124,9 @@ async fn disable_and_enable_all_tracks_in_media_manager() {
     assert!(audio_track.is_muted());
     assert!(!video_track.is_muted());
 
-    video_track.mute_state_transition_to(StableMuteState::Muted);
+    video_track
+        .mute_state_transition_to(StableMuteState::Muted)
+        .unwrap();
     media_connections
         .update_senders(vec![TrackPatch {
             id: video_track_id,
@@ -130,7 +136,9 @@ async fn disable_and_enable_all_tracks_in_media_manager() {
     assert!(audio_track.is_muted());
     assert!(video_track.is_muted());
 
-    audio_track.mute_state_transition_to(StableMuteState::NotMuted);
+    audio_track
+        .mute_state_transition_to(StableMuteState::NotMuted)
+        .unwrap();
     media_connections
         .update_senders(vec![TrackPatch {
             id: audio_track_id,
@@ -140,7 +148,9 @@ async fn disable_and_enable_all_tracks_in_media_manager() {
     assert!(!audio_track.is_muted());
     assert!(video_track.is_muted());
 
-    video_track.mute_state_transition_to(StableMuteState::NotMuted);
+    video_track
+        .mute_state_transition_to(StableMuteState::NotMuted)
+        .unwrap();
     media_connections
         .update_senders(vec![TrackPatch {
             id: video_track_id,