+++ src/signalling/room/mod.rs
@@ -14,8 +14,8 @@ use actix::{
 };
 use derive_more::{Display, From};
 use failure::Fail;
-use medea_client_api_proto::{Event, NegotiationRole, PeerId};
 use futures::future;
+use medea_client_api_proto::{Event, NegotiationRole, PeerId};
 
 use crate::{
     api::control::{